# docs/swagger.yaml
definitions:
...
  responses.BadRequestResponseSwagger:
    properties:
      message:
        description: ข้อความสถานะของ response
        example: Bad Request
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 400
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: false
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
...
  responses.InternalServerErrorResponseSwagger:
    properties:
      message:
        description: ข้อความสถานะของ response
        example: Internal Server Error
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 500
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: false
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
  responses.Response:
    properties:
      data:
        x-order: "4"
      message:
        type: string
        x-order: "3"
      status:
        type: integer
        x-order: "1"
      success:
        type: boolean
        x-order: "0"
      timestamp:
        type: string
        x-order: "2"
    type: object
...
  responses.SuccessResponseSwagger-sim_EZWGetSimGeneralResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/sim.EZWGetSimGeneralResponse'
        description: ข้อมูลที่ต้องการส่งกลับ
        x-order: "4"
      message:
        description: ข้อความสถานะของ response
        example: OK
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 200
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: true
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
  responses.SuccessResponseSwagger-sim_EZWInsertSimRequest:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/sim.EZWInsertSimRequest'
        description: ข้อมูลที่ต้องการส่งกลับ
        x-order: "4"
      message:
        description: ข้อความสถานะของ response
        example: OK
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 200
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: true
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
  responses.SuccessResponseSwagger-sim_EZWSearchSimResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/sim.EZWSearchSimResponse'
        description: ข้อมูลที่ต้องการส่งกลับ
        x-order: "4"
      message:
        description: ข้อความสถานะของ response
        example: OK
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 200
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: true
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
  responses.SuccessResponseSwagger-sim_EZWSimOperatorModelResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/sim.EZWSimOperatorModelResponse'
        description: ข้อมูลที่ต้องการส่งกลับ
        x-order: "4"
      message:
        description: ข้อความสถานะของ response
        example: OK
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 200
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: true
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
  responses.SuccessResponseSwagger-sim_EZWSimStatusModelResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/sim.EZWSimStatusModelResponse'
        description: ข้อมูลที่ต้องการส่งกลับ
        x-order: "4"
      message:
        description: ข้อความสถานะของ response
        example: OK
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 200
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: true
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
  responses.SuccessResponseSwagger-sim_EZWUpdateSimResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/sim.EZWUpdateSimResponse'
        description: ข้อมูลที่ต้องการส่งกลับ
        x-order: "4"
      message:
        description: ข้อความสถานะของ response
        example: OK
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 200
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: true
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
...
  responses.UnauthorizedResponseSwagger:
    properties:
      message:
        description: ข้อความสถานะของ response
        example: Unauthorized
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 401
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: false
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
  sim.EZWGetSimGeneralResponse:
    properties:
      operator_id:
        description: ไอดีผู้ให้บริการซิมการ์ด
        type: integer
        x-order: "2"
      operator_name:
        description: รหัสผู้ให้บริการซิมการ์ด
        type: string
        x-order: "3"
      remark:
        description: หมายเลขสถานะซิมการ์ด
        type: string
        x-order: "6"
      sim_id:
        description: ไอดีของซิมการ์ด
        type: integer
        x-order: "0"
      sim_no:
        description: หมายเลขซิมการ์ด
        type: string
        x-order: "1"
      status_id:
        description: ไอดีสถานะซิมการ์ด
        type: integer
        x-order: "4"
      status_name:
        description: รหัสสถานะซิมการ์ด
        type: string
        x-order: "5"
    type: object
  sim.EZWInsertSimRequest:
    properties:
      operator_id:
        description: |-
          ไอดีผู้ให้บริการซิม
          required: true
        example: 125002
        type: integer
        x-order: "1"
      remark:
        description: |-
          หมายเหตุ
          required: true
        example: สำหรับทดสอบใน API
        type: string
        x-order: "3"
      sim_no:
        description: |-
          เลขที่ซิมการ์ด
          required: true
        example: "66672455949"
        type: string
        x-order: "0"
      status_id:
        description: |-
          ไอดีสถานะซิม
          required: true
        example: 126002
        type: integer
        x-order: "2"
    type: object
  sim.EZWSearchSimRequest:
    properties:
      operator_id:
        description: |-
          ไอดีผู้ให้บริการซิม
          required: true
        example: 99
        type: integer
        x-order: "1"
      remark:
        description: |-
          หมายเหตุ
          required: true
        example: eZView PLUS
        type: string
        x-order: "3"
      sim_no:
        description: |-
          หมายเลขซิมการ์ด
          required: true
        example: "66672455949"
        type: string
        x-order: "0"
      status_id:
        description: |-
          ไอดีสถานะซิม
          required: true
        example: 99
        type: integer
        x-order: "2"
    type: object
  sim.EZWSearchSimResponse:
    properties:
      operator_code:
        description: รหัสผู้ให้บริการ
        example: TrueMove
        type: string
        x-order: "6"
      operator_id:
        description: ไอดีผู้ให้บริการ
        example: 125001
        type: integer
        x-order: "5"
      operator_name:
        description: ชื่อผู้ให้บริการ
        example: TrueMove H
        type: string
        x-order: "7"
      remark:
        description: หมายเหตุ
        example: eZView PLUS
        type: string
        x-order: "8"
      sim_id:
        description: ไอดีของซิมการ์ด
        example: 1
        type: integer
        x-order: "0"
      sim_no:
        description: หมายเลขซิมการ์ด
        example: "66672455949"
        type: string
        x-order: "1"
      status_code:
        description: รหัสสถานะซิมการ์ด
        example: Active
        type: string
        x-order: "3"
      status_id:
        description: ไอดีสถานะซิมการ์ด
        example: 126002
        type: integer
        x-order: "2"
      status_name:
        description: ชื่อสถานะซิมการ์ด
        example: Active
        type: string
        x-order: "4"
    type: object
  sim.EZWSimOperatorModelResponse:
    properties:
      operator_id:
        description: ประเภทไอดีของผู้ให้บริการซิมการ์ด
        example: 125001
        type: integer
        x-order: "0"
      operator_name:
        description: ชื่อประเภทผู้ให้บริการซิมการ์ด
        example: AIS
        type: string
        x-order: "1"
    type: object
  sim.EZWSimStatusModelResponse:
    properties:
      status_id:
        description: ประเภทไอดีของสถานะซิมการ์ด
        example: 126002
        type: integer
        x-order: "0"
      status_name:
        description: ชื่อประเภทสถานะของซิมการ์ด
        example: Inactive
        type: string
        x-order: "1"
    type: object
  sim.EZWUpdateSimRequest:
    properties:
      operator_id:
        description: |-
          ไอดีผู้ให้บริการซิม
          required: true
        example: 125001
        type: integer
        x-order: "2"
      remark:
        description: |-
          หมายเหตุ
          required: true
        example: สำหรับทดสอบใน API
        type: string
        x-order: "4"
      sim_id:
        description: |-
          ไอดีซิมการ์ด
          required: true
        example: 1
        type: integer
        x-order: "0"
      sim_no:
        description: |-
          หมายเลขซิมการ์ด
          required: true
        example: "66672455949"
        type: string
        x-order: "1"
      status_id:
        description: |-
          ไอดีสถานะซิม
          required: true
        example: 126002
        type: integer
        x-order: "3"
    type: object
  sim.EZWUpdateSimResponse:
    properties:
      remark:
        description: |-
          รายละเอียดตอบกลับ
          required: true
        example: eZView PLUS
        type: string
        x-order: "0"
    type: object
...
info:
  contact: {}
  description: For eZVeiw Lite Back Office Project.
  title: eZView Lite Back Office API
  version: "1.0"
paths:
...
  /v1/sim/general:
    get:
      consumes:
      - application/json
      parameters:
      - description: SIM ID
        in: query
        name: sim_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-sim_EZWGetSimGeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: Get SIM general information
      tags:
      - Sim
  /v1/sim/insert:
    post:
      consumes:
      - application/json
      description: เพิ่ม SIM ใหม่ลงฐานข้อมูล
      parameters:
      - description: ข้อมูล SIM ที่ต้องการเพิ่ม
        in: body
        name: sim
        required: true
        schema:
          $ref: '#/definitions/sim.EZWInsertSimRequest'
      produces:
      - application/json
      responses:
        "201":
          description: เพิ่มข้อมูล SIM สำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-sim_EZWInsertSimRequest'
        "400":
          description: ข้อมูลที่ส่งมาไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: เพิ่มข้อมูล SIM
      tags:
      - Sim
  /v1/sim/operator:
    get:
      consumes:
      - application/json
      description: API นี้ใช้สำหรับดึงข้อมูลสถานะซิมการ์ดจากฐานข้อมูล
      produces:
      - application/json
      responses:
        "200":
          description: ดึงข้อมูลสำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-sim_EZWSimOperatorModelResponse'
        "400":
          description: การร้องขอไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: ข้อมูลเข้าสู่ระบบไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: ดึงข้อมูล Sim Operator
      tags:
      - Sim
  /v1/sim/search:
    post:
      consumes:
      - application/json
      description: รับ JSON และส่งพารามิเตอร์ไปยังฟังก์ชัน ezw_search_sim
      parameters:
      - description: Payload สำหรับค้นหา SIM
        in: body
        name: searchRequest
        required: true
        schema:
          $ref: '#/definitions/sim.EZWSearchSimRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ผลลัพธ์การค้นหา SIM
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-sim_EZWSearchSimResponse'
        "400":
          description: ข้อมูลที่ส่งมาไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: ข้อมูลเข้าสู่ระบบไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.Response'
      security:
      - Bearer: []
      summary: ค้นหา SIM
      tags:
      - Sim
  /v1/sim/status:
    get:
      consumes:
      - application/json
      description: API นี้ใช้สำหรับดึงข้อมูลสถานะซิมการ์ดจากฐานข้อมูล
      produces:
      - application/json
      responses:
        "200":
          description: ดึงข้อมูลสำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-sim_EZWSimStatusModelResponse'
        "400":
          description: การร้องขอไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: ข้อมูลเข้าสู่ระบบไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: ดึงข้อมูล Sim Status
      tags:
      - Sim
  /v1/sim/update:
    put:
      consumes:
      - application/json
      description: อัปเดตข้อมูล SIM ที่มีอยู่ในฐานข้อมูล
      parameters:
      - description: ข้อมูล SIM ที่ต้องการอัปเดต
        in: body
        name: sim
        required: true
        schema:
          $ref: '#/definitions/sim.EZWUpdateSimRequest'
      produces:
      - application/json
      responses:
        "200":
          description: อัปเดตข้อมูล SIM สำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-sim_EZWUpdateSimResponse'
        "400":
          description: ข้อมูลที่ส่งมาไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: แก้ไขข้อมูล SIM
      tags:
      - Sim
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"